chore: code review 응답 — 매트릭스 race + doc 정확성 + edge-case tests
PR #75 code review 피드백 5건 반영:

I-2 (Important, 실제 race):
WebmInlineWebView.Coordinator.tryRecreateWebView 가 container?.webView 가
이미 있으면 early-return 했는데, pool 은 그 사이 waiters.removeFirst() 로
waiter 를 이미 소비함 — 슬롯 안 채워진 채 waiter 만 drain 되는 stuck-poster
race. acquire(self) 를 무조건 호출해 pool bookkeeping 우선 정착하고,
container?.webView == nil 가드는 attach 단계로 이동.

I-1 (Important, doc 오류):
NetworkImage 의 .purgeable(true) 코멘트가 "NSCache purgeable" 라고 잘못
적었음. 실제는 SDAnimatedImageView 의 clearBufferWhenStopped 매핑 — 애니
가 *멈출 때* (visibility 변화 등) 버퍼 해제. memory-warning 와 무관.
실제 의미로 다시 적음.

I-3 (Important, doc 오해 소지):
MemoryPressureResponder 가 SDImageCache 도 clearMemory 한다고 강조했지만,
SDMemoryCache 가 이미 같은 noti 자체 listen 하고 자동 호출
(SDMemoryCache.m:69) — 본 responder 의 진짜 가치는 URLCache.shared 의 50MB
mem 캐시 flush. doc 정정해서 belt-and-suspenders 임을 명시.

I-5 (Important, 테스트 누락):
- testRespondBeforeInstallDefaultHandlersIsNoOp: AppDelegate 미실행 환경
  에서 respond() 호출이 crash 안 함 (default no-op closure)
- testReleaseOnNeverAcquiredHolderIsNoOp: dismantleUIView 외 경로
  (acquire 전 사라지는 race) 에서 release 무해

M-1 (Minor):
NetworkImage 의 .maxBufferSize(UInt(16*1024*1024)) — UInt cast 불필요.
정수 리터럴이 자동 추론. 제거.

defer/skip:
- I-4 (compactDeadRefs 빈도): 실제 영향 낮음, SwiftUI dismantleUIView
  신뢰성 OK
- M-3 (maxConcurrent 가변): out of scope
- M-4 (constraint deactivate): 1회용 container 라 zombie 없음
- M-5 (DEBUG print spam): 별도 cleanup PR
- M-6 (gateOnVisibility async revisit): 사용자 직접 요청한 라인이라 유지

241 → 243 tests (+2 신규 edge case), 모두 PASS.

+++ nunting/Views/NetworkImage.swift
@@ -161,10 +161,16 @@ struct NetworkImage: View {
                 // 800×500 — enough to keep the visible loop smooth, while
                 // forcing re-decode on long animations rather than holding
                 // every frame in RAM forever.
-                .maxBufferSize(UInt(16 * 1024 * 1024))
-                // Mark the decoded frame buffer as `NSCache`-purgeable so
-                // a memory pressure event evicts frames eagerly instead
-                // of waiting for the cap.
+                .maxBufferSize(16 * 1024 * 1024)
+                // SDWebImageSwiftUI 의 `.purgeable(true)` 는 NSCache 의
+                // purgeable 플래그가 아니라 SDAnimatedImageView 의
+                // `clearBufferWhenStopped` 로 매핑됨 (AnimatedImage.swift:693).
+                // 의미: 애니메이션이 *멈출 때* 디코드된 프레임 버퍼 해제.
+                // LazyVStack 스크롤 재활용으로 off-screen 됐을 때
+                // visibility 변화 → 정지 → 버퍼 해제 경로가 발화 — 본문
+                // 짤방이 화면 밖에 오랫동안 남아있는 메모리 잔존을 줄임.
+                // memory-warning 와는 무관 (그건 SDImageCache 가 자체
+                // 처리).
                 .purgeable(true)
                 .resizable()
                 .scaledToFit()
